Alteração no código dos botões de deletar dos formulários de cadastros de usuário, jogador e compra

+++ src/view/JFrmCadUsuario.java
@@ -12,6 +12,7 @@ import java.util.List;
 import javax.persistence.RollbackException;
 import javax.swing.ImageIcon;
 import javax.swing.JFrame;
+import javax.swing.JOptionPane;
 import javax.swing.JPanel;
 
 /**
@@ -19,7 +20,7 @@ import javax.swing.JPanel;
  * @author Arthur Oliveira
  */
 public class JFrmCadUsuario extends JPanel {
-    
+
     public JFrmCadUsuario() {
         initComponents();
         if (!Beans.isDesignTime()) {
@@ -245,7 +246,6 @@ public class JFrmCadUsuario extends JPanel {
         }
     }// </editor-fold>//GEN-END:initComponents
 
-    
     @SuppressWarnings("unchecked")
     private void refreshButtonActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_refreshButtonActionPerformed
         entityManager.getTransaction().rollback();
@@ -260,13 +260,19 @@ public class JFrmCadUsuario extends JPanel {
 
     private void deleteButtonActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_deleteButtonActionPerformed
         int[] selected = masterTable.getSelectedRows();
-        List<view.Usuario> toRemove = new ArrayList<view.Usuario>(selected.length);
-        for (int idx = 0; idx < selected.length; idx++) {
-            view.Usuario u = list.get(masterTable.convertRowIndexToModel(selected[idx]));
-            toRemove.add(u);
-            entityManager.remove(u);
+        if (selected.length > 0) {
+            int r = JOptionPane.showConfirmDialog(null, "Deseja excluir o registro?", "Confirmação", JOptionPane.YES_NO_OPTION);
+            if (r == 0) {
+                List<view.Usuario> toRemove = new ArrayList<view.Usuario>(selected.length);
+                for (int idx = 0; idx < selected.length; idx++) {
+                    view.Usuario u = list.get(masterTable.convertRowIndexToModel(selected[idx]));
+                    toRemove.add(u);
+                    entityManager.remove(u);
+                }
+                list.removeAll(toRemove);
+                saveButton.doClick();
+            }
         }
-        list.removeAll(toRemove);
     }//GEN-LAST:event_deleteButtonActionPerformed
 
     private void newButtonActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_newButtonActionPerformed
@@ -277,7 +283,7 @@ public class JFrmCadUsuario extends JPanel {
         masterTable.setRowSelectionInterval(row, row);
         masterTable.scrollRectToVisible(masterTable.getCellRect(row, 0, true));
     }//GEN-LAST:event_newButtonActionPerformed
-    
+
     private void saveButtonActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_saveButtonActionPerformed
         try {
             entityManager.getTransaction().commit();
@@ -355,5 +361,5 @@ public class JFrmCadUsuario extends JPanel {
             }
         });
     }
-    
+
 }
